🎨  refactor : 오타 수정 및 폴더 구조 변경
