+++ src/components/question/ProfileSection.jsx
@@ -1,5 +1,5 @@
 import ProfileImg from '../../assets/images/profile_img.svg?react';
-import logo from '../../assets/images/logo.svg';
+import Logo from '../../assets/images/logo.svg?react';
 import LinkImg from '../../assets/images/Link.svg?react';
 import KakaoImg from '../../assets/images/Kakaotalk.svg?react';
 import FacebookImg from '../../assets/images/Facebook.svg?react';
@@ -7,19 +7,19 @@ import FacebookImg from '../../assets/images/Facebook.svg?react';
 function ProfileSection() {
   return (
     <section className="absolute top-[50px] flex justify-center flex-col items-center gap-[12px]">
-      <img className="" src={logo} alt="로고 이미지" />
-      <ProfileImg className="w-[136px]" />
+      <Logo alt="로고 이미지" />
+      <ProfileImg className="w-[136px] h-[136px]" />
       <p className="text-[32px] ">아초는고양이</p>
       <div className="flex gap-[12px]">
-        <div className="p-[11px] bg-bn-40 flex justify-center items-center rounded-full">
+        <button className="p-[11px] bg-bn-40 flex justify-center items-center rounded-full">
           <LinkImg className="fill-gs-10" />
-        </div>
-        <div className="p-[11px] bg-y50 flex justify-center items-center rounded-full">
+        </button>
+        <button className="p-[11px] bg-y50 flex justify-center items-center rounded-full">
           <KakaoImg className="fill-gs-60" />
-        </div>
-        <div className="p-[11px] bg-b50 flex justify-center items-center rounded-full">
+        </button>
+        <button className="p-[11px] bg-b50 flex justify-center items-center rounded-full">
           <FacebookImg className="fill-gs-10" />
-        </div>
+        </button>
       </div>
     </section>
   );
